// src/extensions/tsserver/protocol.d.ts
  interface FileRequestArgs {
    /**
     * The file for the request (absolute pathname required).
     */
    file: string;
    projectFileName?: string;
  }
// ...
  interface Location {
    line: number;
    offset: number;
  }
  /**
   * Object found in response messages defining a span of text in source code.
   */
  interface TextSpan {
    /**
     * First character of the definition.
     */
    start: Location;
    /**
     * One character past last character of the definition.
     */
    end: Location;
  }
// ...
  interface NavigationBarItem {
    /**
     * The item's display text.
     */
    text: string;
    /**
     * The symbol's kind (such as 'className' or 'parameterName').
     */
    kind: ScriptElementKind;
    /**
     * Optional modifiers for the kind (such as 'public').
     */
    kindModifiers?: string;
    /**
     * The definition locations of the item.
     */
    spans: TextSpan[];
    /**
     * Optional children.
     */
    childItems?: NavigationBarItem[];
    /**
     * Number of levels deep this item should appear.
     */
    indent: number;
  }
  /** protocol.NavigationTree is identical to ts.NavigationTree, except using protocol.TextSpan instead of ts.TextSpan */
  interface NavigationTree {
    text: string;
    kind: ScriptElementKind;
    kindModifiers: string;
    spans: TextSpan[];
    nameSpan: TextSpan | undefined;
    childItems?: NavigationTree[];
  }